Exercise 1
Create two new activities (Activity3 and Activity4) with different
layouts and views
(student’s choice). In the first activity (MainActivity) include a new
button that
shows Activity3 when it is clicked. In Activity3 create another button
for showing
Activity4. Finally, when Activity4 is closed, show directly the first
activity
(MainActivity).